ft_split finally

+++ ft_putchar_fd.c
@@ -6,7 +6,7 @@
 /*   By: mfassbin <mfassbin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 16:53:46 by mfassbin          #+#    #+#             */
-/*   Updated: 2023/10/13 17:01:56 by mfassbin         ###   ########.fr       */
+/*   Updated: 2023/10/14 22:44:52 by mfassbin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putchar_fd(char c, int fd)
 {
-	write(fd, &c, 1);
+	write((unsigned int)fd, &c, 1);
 }
 /* 
 int main(void)
