+++ ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: mfassbin <mfassbin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 17:02:47 by mfassbin          #+#    #+#             */
-/*   Updated: 2023/10/13 17:19:13 by mfassbin         ###   ########.fr       */
+/*   Updated: 2023/10/14 22:45:01 by mfassbin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	ft_putendl_fd(char *s, int fd)
 {
 	while (*s)
 	{
-		write(fd, s, 1);
+		write((unsigned int)fd, s, 1);
 		s++;
 	}
 	write(fd, "\n", 1);
