+++ ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: mfassbin <mfassbin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 17:02:47 by mfassbin          #+#    #+#             */
-/*   Updated: 2023/10/13 17:11:46 by mfassbin         ###   ########.fr       */
+/*   Updated: 2023/10/14 22:32:23 by mfassbin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	ft_putstr_fd(char *s, int fd)
 {
 	while (*s)
 	{
-		write(fd, s, 1);
+		write((unsigned int)fd, s, 1);
 		s++;
 	}
 }
